refactor: 두두 채팅 관련 도메인 로직을 AI 모듈에서 분리 (ai -> chat)
