+++ apps/backend/src/common/docs/openapi.ts
@@ -1,6 +1,5 @@
 import { OpenAPIRegistry, OpenApiGeneratorV3 } from '@asteasolutions/zod-to-openapi';
 
-import { registerAIApi } from 'src/features/ai/ai.docs';
 import { registerCommonSchemas } from './common.docs';
 import { registerBehaviorApi } from '../../features/behavior/behavior.docs';
 import { registerGoalApi } from '../../features/goal/goal.docs';
@@ -17,7 +16,6 @@ registerGoalApi(registry);
 registerBehaviorApi(registry, commonSchemas);
 registerAuthApi(registry, commonSchemas);
 registerStatApi(registry, commonSchemas);
-registerAIApi(registry, commonSchemas);
 registerPushApi(registry, commonSchemas);
 
 type OpenApiDocument = ReturnType<OpenApiGeneratorV3['generateDocument']>;
