+++ apps/backend/src/features/push/push.module.ts
@@ -4,7 +4,7 @@ import { PushController } from './push.controller';
 import { PushService } from './push.service';
 import { PushSubscriptionEntity } from './push-subscription.entity';
 import { User } from '../user/user.entity';
-import { DodoChatMessage } from '../ai/dodo-chat-message.entity';
+import { DodoChatMessage } from '../chat/dodo-chat-message.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([PushSubscriptionEntity, User, DodoChatMessage])],
