+++ packages/shared/src/schemas/chat.schemas.ts
