+++ packages/shared/src/schemas/index.ts
@@ -4,5 +4,5 @@ export * from './behavior.schemas';
 export * from './user.schemas';
 export * from './common.schemas';
 export * from './stat.schemas';
-export * from './ai.schemas';
+export * from './chat.schemas';
 export * from './push.schemas';
